修复64位编译问题

+++ interfaces/innerkits/accesstoken/include/access_token.h
@@ -16,11 +16,13 @@
 #ifndef ACCESS_TOKEN_DEF_H
 #define ACCESS_TOKEN_DEF_H
 
+#include <string>
+
 namespace OHOS {
 namespace Security {
 namespace AccessToken {
 typedef unsigned int AccessTokenID;
-typedef unsigned long long FullTokenID;
+typedef uint64_t FullTokenID;
 typedef unsigned int AccessTokenAttr;
 static const int DEFAULT_TOKEN_VERSION = 1;
 static const AccessTokenID INVALID_TOKENID = 0;
